Artist Model Modified-Mus Dir, lyrsct,Sngr merged

+++ resources/views/artists/index.blade.php
@@ -11,7 +11,7 @@
                       <i class="fa fa-dashboard"></i>  <a href="{{ url('/admin/songs')}}">Dashboard</a>
                   </li>
                   <li class="active ml-3">
-                      <i class="fa fa-table"></i> All Artists
+                      <i class="fa fa-table"></i> {{ $title ?? 'Artists' }}
                   </li>
               </ol>
           </div>
@@ -21,12 +21,12 @@
       <div class="alert alert-success">
           <p>{{ $message }}</p>
       </div>
-    @endif
+    @endif   
 
     <div class="row mb-4">
         <div class=" d-flex justify-content-between">
             <div class="">
-                <h3>All Artists</h3>
+                <h3>All  {{ $title ?? 'Artists' }}</h3>
             </div>
             <div class="">  
                 <a class="btn btn-primary ml-auto" href="{{ url('/admin/artists/create')}}">Add Artist</a>
@@ -42,16 +42,22 @@
                     <div class="col-lg-3 mb-4">
                         <div class="bg-column" style="background-image: url('{{ URL::to('/uploads/artists/' .  $value->image_path)  }}');">
 
-                            <div class="bg-title">{{ $value->name }}</div>
+                       
+                            <a href="{{route('artists.show', $value->id)}}" target="_self">
+                                <div class="bg-title">{{ $value->name }}</div>      
+                            </a>
 
                         </div>
-
+                        
                         <div class="bg-options text-center mt-2">    
                             <a class="btn btn-primary btn-sm" href="{{route('artists.edit', $value->id)}}"><i class="fa fa-pencil" aria-hidden="true"></i></a>
-                            {{ Form::open(['method' => 'DELETE','route' => ['artists.destroy', $value->id],'style'=>'display:inline']) }}
-                            <button type="submit" style="display: inline;" class="btn btn-danger btn-sm show_confirm"><i class="fa fa-trash" aria-hidden="true"></i></button>
-                            {{ Form::close() }}
+                            @if ( $allow_delete  )
+                                {{ Form::open(['method' => 'DELETE','route' => ['artists.destroy', $value->id],'style'=>'display:inline']) }}
+                                <button type="submit" style="display: inline;" class="btn btn-danger btn-sm show_confirm"><i class="fa fa-trash" aria-hidden="true"></i></button>
+                                {{ Form::close() }}
+                            @endif
                         </div>
+                        
                     </div>
                     @endforeach                      
               </div>
